Refactor proposal management and enhance user permissions
- Updated the proposal editing page to initialize the Project class with form data.
- Improved project type selection logic based on the collection type.
- Enhanced status display and added warnings for status changes in proposals.
- Refactored applicant and author management UI for better usability.
- Added scholarship and university fields with modal search functionality.
- Updated project and proposal listing pages to include filtering options by type.
- Implemented permission checks for viewing and editing proposals.
- Added new roles for editing proposals in roles.json.
- Improved API filtering to restrict access based on user permissions.
- Refactored organization icon retrieval in Organization.php.
- Updated Project.php to handle new fields related to scholarships and universities.
- Adjusted migration scripts to handle stipendiate projects.

+++ routes/migrate.php
@@ -33,6 +33,18 @@ Route::get('/migrate/test', function () {
 
     include_once BASEPATH . "/php/Project.php";
     $Project = new Project;
+
+
+    // first check if there are stipendiates. If yes, we need to migrate them into a new project type
+    $stipendiates = $osiris->projects->find(['type' => 'Stipendium'])->toArray();
+    if (count($stipendiates) > 0) {
+        echo "<p>There are ".count($stipendiates)." stipendiate projects. Please migrate them to the new activity type.</p>";
+        foreach ($stipendiates as $doc) {
+            dump($doc['status']);
+        }
+        die;
+    }
+
     // Drittmittel
     $osiris->adminProjects->deleteOne(['id' => 'third-party']);
     $osiris->adminProjects->insertOne([
@@ -116,32 +128,32 @@ Route::get('/migrate/test', function () {
         ['$set' => ['type' => 'third-party']]
     );
 
-    $osiris->adminProjects->deleteOne(['id' => 'stipendate']);
-    $osiris->adminProjects->insertOne([
-        'id' => 'stipendate',
-        'icon' => 'tip-jar',
-        'color' => '#63a308',
-        'name' => 'Scholarship',
-        'name_de' => 'Stipendium',
-        'modules' => [
-            'abstract',
-            'public',
-            'internal_number',
-            'website',
-            'grant_sum',
-            'supervisor',
-            'scholar',
-            'scholarship',
-            'university',
-        ],
-        'topics' => false,
-        'disabled' => false,
-        'portfolio' => true
-    ]);
-    $osiris->projects->updateMany(
-        ['type' => 'Stipendium'],
-        ['$set' => ['type' => 'stipendiate']]
-    );
+    // $osiris->adminProjects->deleteOne(['id' => 'stipendate']);
+    // $osiris->adminProjects->insertOne([
+    //     'id' => 'stipendate',
+    //     'icon' => 'tip-jar',
+    //     'color' => '#63a308',
+    //     'name' => 'Scholarship',
+    //     'name_de' => 'Stipendium',
+    //     'modules' => [
+    //         'abstract',
+    //         'public',
+    //         'internal_number',
+    //         'website',
+    //         'grant_sum',
+    //         'supervisor',
+    //         'scholar',
+    //         'scholarship',
+    //         'university',
+    //     ],
+    //     'topics' => false,
+    //     'disabled' => false,
+    //     'portfolio' => true
+    // ]);
+    // $osiris->projects->updateMany(
+    //     ['type' => 'Stipendium'],
+    //     ['$set' => ['type' => 'stipendiate']]
+    // );
 
     $osiris->adminProjects->deleteOne(['id' => 'subproject']);
     $osiris->projects->updateMany(
